requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
uvicorn==0.30.1
fastapi==0.111.0